+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.0.4
+Added option to suppress the ascii output and -h to show help.
+
+
+# 1.0.3
+* Exposed hexdump as a executable.
+* upgraded to dcli 1.8
+* Added search for differnt file extensions as windows users expect.
+
 # 1.0.2
 
 # 1.0.1